Fix Loading Issues when DeluxeChat or PlaceholderAPI were not installed
