+++ src/main/java/denniss17/dsTitle/DSTitle.java
@@ -14,12 +14,6 @@ import com.kaltiz.dsTitle.storage.SQLTitleStorage;
 import com.kaltiz.dsTitle.storage.TitleStorage;
 import com.kaltiz.dsTitle.storage.YMLTitleStorage;
 
-import me.clip.deluxechat.placeholders.DeluxePlaceholderHook;
-import me.clip.deluxechat.placeholders.PlaceholderHandler;
-import me.clip.placeholderapi.PlaceholderAPI;
-import me.clip.placeholderapi.PlaceholderHook;
-
-
 public class DSTitle extends JavaPlugin{	
 	private PermissionManager permissionManager;
 	private TeamManager teamManager;
@@ -62,14 +56,14 @@ public class DSTitle extends JavaPlugin{
         }
 		if (Bukkit.getPluginManager().isPluginEnabled("PlaceholderAPI"))
 		{
-			if (RegisterPlaceHolderHooks(title)) {
+			if (Placeholders.RegisterPlaceHolderHooks(title)) {
 		        getLogger().info("dsTitle was successfully registered with PlaceHolderAPI!");
 		        placeHolders = true;
 		    }
 		}
 		if(Bukkit.getPluginManager().isPluginEnabled("DeluxeChat"))
 		{
-			if (RegisterDeluxeChatHooks(title)) {
+			if (Placeholders.RegisterDeluxeChatHooks(title)) {
 		        getLogger().info("dsTitle was successfully registered with DeluxeChat!");
 		        placeHolders = true;
 		    }
@@ -169,104 +163,5 @@ public class DSTitle extends JavaPlugin{
 		receiver.sendMessage(ChatStyler.setTotalStyle(message));
 	}
 	
-	/**
-	 * Registers Placeholders into PlaceholderAPI for dsTitle Suffixes/Prefixes
-	 * Use these in PlaceholderAPI by using %dsTitle_prefix% for prefix or %dsTitle_suffix% for suffix
-	 * @return True if placeholders were registered successfully, false if failed
-	 */
-	public boolean RegisterPlaceHolderHooks(final DSTitle plugin){
-		boolean hooked = PlaceholderAPI.registerPlaceholderHook(plugin, new PlaceholderHook(){
-			public String onPlaceholderRequest(Player p, String identifier) {
-				if(identifier.equals("prefix")){
-                	TitleManager man = plugin.getTitleManager();
-                    if (man != null)
-                    {
-                  	String prefixTag;
-                  	prefixTag = man.getPrefixChatTag(p);
-                      if (prefixTag != null)
-                      {
-                        return man.getPrefixChatTag(p);
-                      }else
-                      {
-                        return " ";
-                      }
-                    }else
-                    {
-                      return " ";
-                    }
-				}
-				if(identifier.equals("suffix")){
-                	TitleManager man = plugin.getTitleManager();
-                    if (man != null)
-                    {
-                  	String suffixTag;
-                  	suffixTag = man.getSuffixChatTag(p);
-                      if (suffixTag != null)
-                      {
-                        return man.getSuffixChatTag(p);
-                      }else
-                      {
-                        return " ";
-                      }
-                    }else
-                    {
-                      return " ";
-                    }
-				}
-				return null;
-			}			
-		});
-		return hooked;
-	}
 	
-	/**
-	 * Registers Placeholders into DeluxeChat for dsTitle Suffixes/Prefixes
-	 * Use these in DeluxeChat by using %dsTitle_prefix% for prefix or %dsTitle_suffix% for suffix
-	 * @return True if placeholders were registered successfully, false if failed
-	 */
-	public boolean RegisterDeluxeChatHooks(final DSTitle plugin){
-		boolean hookedchat = PlaceholderHandler.registerPlaceholderHook(plugin, new DeluxePlaceholderHook(){
-			public String onPlaceholderRequest(Player p, String identifier) {
-                if(identifier.equals("prefix")){
-                	TitleManager man = plugin.getTitleManager();
-                    if (man != null)
-                    {
-                  	String prefixTag;
-                  	prefixTag = man.getPrefixChatTag(p);
-                      if (prefixTag != null)
-                      {
-                        return man.getPrefixChatTag(p);
-                      }else
-                      {
-                        return " ";
-                      }
-                    }else
-                    {
-                      return " ";
-                    }
-				}
-                if(identifier.equals("suffix")){
-                	TitleManager man = plugin.getTitleManager();
-                    if (man != null)
-                    {
-                  	String suffixTag;
-                  	suffixTag = man.getSuffixChatTag(p);
-                      if (suffixTag != null)
-                      {
-                        return man.getSuffixChatTag(p);
-                      }else
-                      {
-                        return " ";
-                      }
-                    }else
-                    {
-                      return " ";
-                    }
-				}
-				return null;
-			}
-			
-		});
-		return hookedchat;
-	}
 }
